requirements.txt
Django
django-cors-middleware
djangorestframework
djangorestframework-jwt
serpy
